feat(web): the queue is one day at a time, and the calendar moves beside the date
An experiment, and two changes that turned out to be one.

**The calendar left the rail.** It stood open in 288 px of every desktop
view to answer a question a reader asks rarely, and it did not exist at
all below 1024 px. It is now behind `DateFilter`, beside the heading:
a popover on a desktop, the bottom sheet it already used on a phone.
That is where `design.md` §3 named the date picker in the first place —
the rail was the departure, and the desktop popover was never built.

**The queue is a day.** `/queue` opens on the newest day still holding
something and holds it; `/queue/2026-09-20` is one day. A day can be
finished and a river cannot, so the screen a reader lands on has an end
to it — and clearing one now says so, names the next day and how much is
on it, and offers to open it. There is deliberately no way back to the
whole list: an experiment with the old shape one click away only
measures the click.

Which days those are is its own query, `useUnreadDays`, walking the
digest's compact rows to a null cursor. The day rail knew only the days
already loaded; this knows the backlog, and knowing it means a step is a
range query for one day rather than a walk forward through everything in
between.

The two screens are one shape now — a date, a picker, rows — so each
says in one line what it holds that the other does not, and both
calendars count the same thing through `lib/day-counts.ts`. The queue's
cells came from its unread index before, so a day read to the end had no
cell and the picker said nothing arrived when seven summaries had.

Smaller, in passing: the calendar's window abbreviates its months, since
a five-week window straddling two of them put two full names and a year
across a 288 px control. The queue's ending no longer counts the rows on
screen — day-scoped, "that is all 3 unread summaries" was true of the
screen and false about the reader.

+++ apps/web/src/lib/day.ts
@@ -134,7 +134,13 @@ export function shiftAnchor(anchor: DayKey, windows: number): DayKey {
   return dayKeyOf(date.getTime());
 }
 
-/** What a window of days is called: one month, or the two it straddles. */
+/**
+ * What a window of days is called: one month, or the two it straddles. Abbreviated — "Sep 2026",
+ * "Aug – Sep 2026" — because a five-week window straddling two months put two full month names and
+ * a year across the top of a 288 px control (owner decision 2026-09-22). Still the month by name,
+ * never a number, and still the reader's own language: `short` is the locale's abbreviation, not a
+ * three-letter slice of ours. A cell's accessible name keeps the full month (`fullDate`).
+ */
 export function windowLabel(days: readonly DayKey[]): string {
   const first = days[0];
   const last = days[days.length - 1];
@@ -143,7 +149,7 @@ export function windowLabel(days: readonly DayKey[]): string {
   const to = new Date(dayBounds(last).fromMs);
   const month = (date: Date, withYear: boolean) =>
     date.toLocaleDateString(undefined, {
-      month: "long",
+      month: "short",
       ...(withYear ? { year: "numeric" } : {}),
     });
   if (
